feat(project): make get by tag test unconditional
Before this commit, the test would pass even if

```
/articles?tag=dragons
```

returned empty, and I had a broken functionality in my implementation
unnoticed due to this.

However, this should not be possible, since the earlier Create Article
test creates an article with the tag dragons:

```
\"tagList\":[\"dragons\",\"training\"]
```

Presumably this check exists to help the test pass on the upstream API
<https://conduit.productionready.io/api>, however the only way this could
fail even there would be for the data to fall off the 500 article cutoff,
which is unlikely during the runtime of a test.